+++ package.json
@@ -5,7 +5,9 @@
   "main": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon src/index.js"
+    "startOld": "nodemon src/index.js",
+    "build": "tsc --project ./",
+    "start": " ts-node-dev --project ./tsconfig.json --respawn ./src/index.ts"
   },
   "keywords": [],
   "author": "",
@@ -22,10 +24,12 @@
     "json2csv": "^5.0.6",
     "moment": "^2.29.1",
     "swagger-ui-express": "^4.1.6",
+    "ts-node-dev": "^2.0.0",
     "underscore": "^1.13.1",
     "web3": "^1.6.1"
   },
   "devDependencies": {
+    "@types/express": "^4.17.13",
     "nodemon": "^2.0.13"
   }
 }
