+++ src/utils.ts
@@ -1,19 +1,24 @@
+import {DonationResponse} from "./types/general";
+
 const {ethers} = require("ethers");
 const Web3 = require('web3');
 const {pinJSONToIPFS} = require("./pinataUtils");
 const {hexlify, solidityKeccak256} = ethers.utils;
 
 
-const createSmartContractCallAddBatchParams = async ({
-                                                       nrGIVAddress,
-                                                       donationsWithShare,
-                                                       givRelayerAddress
-                                                     }, maxAddressesPerFunctionCall) => {
+export const createSmartContractCallAddBatchParams = async (params: {
+                                                       nrGIVAddress: string,
+                                                       donationsWithShare:DonationResponse[] ,
+                                                       givRelayerAddress: string
+                                                     }, maxAddressesPerFunctionCall:number) => {
+  const { nrGIVAddress,
+    donationsWithShare,
+    givRelayerAddress} = params;
   const partNumbers = donationsWithShare.length / maxAddressesPerFunctionCall
   let result = `connect ${nrGIVAddress} token-manager voting:1 act agent:0 ${givRelayerAddress} `;
   result += 'addBatches(bytes32[],bytes) ['
-  const hashParams = {
-    ipfsLink:''
+  const hashParams :any= {
+    ipfsLink:'',
   }
   let nonce = await getLastNonceForWalletAddress(givRelayerAddress)
   const rawDatasForHash  = []
@@ -46,10 +51,14 @@ const createSmartContractCallAddBatchParams = async ({
 
 
 
-const getSmartContractAddBatchesHash =  ({
-                                          donationsWithShare,
-                                          nonce
+const getSmartContractAddBatchesHash =  (params: {
+                                          donationsWithShare: DonationResponse[],
+                                          nonce :number
                                         }) => {
+  const {
+    donationsWithShare,
+        nonce
+  } = params
   const rawData =  {
     nonce,
     amounts: donationsWithShare.map(
@@ -65,7 +74,7 @@ const getSmartContractAddBatchesHash =  ({
 }
 
 
-const convertExponentialNumber = (n) => {
+const convertExponentialNumber = (n:number) => {
   const sign = +n < 0 ? "-" : "",
     toStr = n.toString();
   if (!/e/i.test(toStr)) {
@@ -75,17 +84,22 @@ const convertExponentialNumber = (n) => {
     .replace(/^-/, "")
     .replace(/^([0-9]+)(e.*)/, "$1.$2")
     .split(/e|\./);
-  return +pow < 0
-    ? sign + "0." + "0".repeat(Math.max(Math.abs(pow) - 1 || 0, 0)) + lead + decimal
+  return Number(pow) < 0
+    ? sign + "0." + "0".repeat(Math.max(Math.abs(Number(pow)) - 1 || 0, 0)) + lead + decimal
     : sign + lead + (+pow >= decimal.length ? (decimal + "0".repeat(Math.max(+pow - decimal.length || 0, 0))) : (decimal.slice(0, +pow) + "." + decimal.slice(+pow)))
 }
 
 
-function hashBatchEthers({
-                           nonce,
-                           recipients,
-                           amounts
+function hashBatchEthers(params: {
+                           nonce : number,
+                           recipients: string[],
+                           amounts: string[]
                          }) {
+  const {
+    nonce,
+    recipients,
+    amounts
+  } = params
   console.log('hashBatchEthers() input', {nonce, recipients, amounts})
   const hash = hexlify(
     solidityKeccak256(
@@ -101,14 +115,10 @@ function hashBatchEthers({
 const xdaiWeb3NodeUrl = process.env.XDAI_NODE_HTTP_URL
 const xdaiWeb3 = new Web3(xdaiWeb3NodeUrl);
 
-const getLastNonceForWalletAddress = async (walletAddress) => {
+export const getLastNonceForWalletAddress = async (walletAddress: string) : Promise<number>=> {
   const userTransactionsCount = await xdaiWeb3.eth.getTransactionCount(
     walletAddress
   );
   return userTransactionsCount - 1
 }
 
-module.exports = {
-  createSmartContractCallAddBatchParams,
-  getLastNonceForWalletAddress
-}
